Update is here!

+++ discord/say.js
@@ -3,6 +3,7 @@ module.exports = {
 	admin: true,
 	pm: true,
 	commandFunction: function (args, message, Bot) {
+		if (message.guild.id === '719076445699440700') return Bot.say(message.channel.name, args.join(' '));
 		message.channel.send(args.join(' '));
 		message.delete();
 	}
